+++ crates/portal-rest/clients/ts/src/index.ts
@@ -60,6 +60,13 @@ export {
   BurnCashuRequest,
   CashuResponseStatus,
 
+  // Verification
+  CreateVerificationSessionRequest,
+  VerificationSessionResponse,
+
+  // Portal Token
+  RequestVerificationTokenRequest,
+
   // Relays
   RelayRequest,